+++ frontend/src/constants/errorMessage.ts
@@ -15,3 +15,5 @@ export const API_ERROR_MESSAGE: ApiErrorMessages = {
 export const SERVER_ERROR_REGEX = /^5\d{2}$/;
 
 export const ROUTE_ERROR_MESSAGE = '찾으시는 페이지가 없어요.';
+
+export const INVALID_GROUP_ACCESS_CODE_MESSAGE = '올바르지 않은 확인 코드예요.';
