+++ frontend/src/mocks/handlers/group.ts
@@ -2,7 +2,7 @@ import { http, HttpResponse } from 'msw';
 
 import endPoint from '@/apis/endpoints';
 
-import { CREATED_GROUP_DATA } from '../mockData/createdGroupData';
+import { CREATED_GROUP_DATA, INVALID_GROUP_ACCESS_CODE } from '../mockData/group';
 
 const postDataForUrl = () => {
   return http.post(endPoint.postingDataForURL, async () => {
@@ -10,13 +10,30 @@ const postDataForUrl = () => {
   });
 };
 
-// NOTE: 에러 테스트용 핸들러
 // const postDataForUrl = () => {
 //   return http.post(endPoint.postingDataForURL, async () => {
 //     return HttpResponse.json({ error: '서버 에러 테스트' }, { status: 500 });
 //   });
 // };
 
-const groupHandler = [postDataForUrl()];
+const getIsValidGroupAccessCode = () => {
+  return http.get(endPoint.gettingReviewList, async () => {
+    return HttpResponse.json({ status: 200 });
+  });
+};
+
+// const getIsValidGroupAccessCode = () => {
+//   return http.get(endPoint.gettingReviewList, async () => {
+//     return HttpResponse.json(INVALID_GROUP_ACCESS_CODE, { status: 400 });
+//   });
+// };
+
+// const getIsValidGroupAccessCode = () => {
+//   return http.get(endPoint.gettingReviewList, async () => {
+//     return HttpResponse.json({ error: '서버 에러 테스트' }, { status: 500 });
+//   });
+// };
+
+const groupHandler = [postDataForUrl(), getIsValidGroupAccessCode()];
 
 export default groupHandler;
