--- frontend/src/mocks/mockData/createdGroupData.ts
@@ -1,4 +1,0 @@
-export const CREATED_GROUP_DATA = {
-  reviewRequestCode: 'mocked-reviewRequestCode',
-  groupAccessCode: 'mocked-groupAccessCode',
-};
